requirements.txt
# Core Dependencies
pinecone-client>=3.0.0
pypdf>=3.17.0
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-text-splitters>=0.0.1
tqdm>=4.66.0
google-generativeai>=0.3.0

# Additional Dependencies (automatically installed)
# numpy>=1.24.0
# torch>=2.0.0
# transformers>=4.30.0
# huggingface-hub>=0.16.0

# Optional for Development
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# python-dotenv>=1.0.0